Attribute Small Changes(double dataType and is_searchable field)

See merge request sunread/sunread-api!178

+++ Modules/Product/Database/Migrations/2021_05_05_110932_create_product_attribute_decimal_table.php
@@ -10,7 +10,7 @@ class CreateProductAttributeDecimalTable extends Migration
     {
         Schema::create("product_attribute_decimal", function (Blueprint $table) {
             $table->id();
-            $table->decimal("value");
+            $table->double("value");
             $table->timestamps();
         });
     }
